+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - CodeCaller value builder (moved from ganbarodigital/php-reflection)
 - ExceptionCaller trait
+- ExceptionMessageData trait (in the Traits folder)
 
 ## 1.0.0 - Sat Jun 27 2015